refactor(settings): split CloudSecretsSettingsSurface into pattern + access hook per D5
CloudSecretsSettingsSurface (product-surfaces) mixed the Cloud secrets
queries/mutations with the panel render in one connected component. Per D5,
splits it: SecretManagementPanel (product-ui/src/secrets/, already
presentational — data in, callbacks out, zero SDK imports) is the pattern
half; the new use-cloud-secrets-panel hook
(product-client/src/hooks/access/cloud/) owns the query/mutation wiring
against @proliferate/cloud-sdk-react and returns SecretManagementPanel's
full prop set, including the scope-driven title/description/filePathMode/
canManage metadata (canManage still ANDs the scope's raw authority with
runtime enabled/query-error state, preserving the original behavior exactly).

Rewires the 3 real call sites to compose hook + pattern directly:
OrganizationSecretsPane, PersonalSecretsPane, RepoEnvironmentPane (the
internal CloudEnvironmentDetail consumer was already deleted earlier in this
cull as part of the dead CloudEnvironmentsSettingsSurface bundle). Deletes
the old CloudSecretsSettingsSurface and its product-surfaces exports entry;
it had no dedicated test to port (untested prior to this split).
